refactor: move volc task delete handler out of graylight package
VolcTaskDelete and its helpers (buildVolcDeleteResp, isTerminalStatus,
volcDeleteMapStatus) previously lived in controller/graylight/, which
the upstream PR cannot depend on. Move them to controller/ proper —
the package was also empty after this move, so the directory is gone.

router/video-router.go now imports only the single controller package
instead of both controller and controller/graylight.

No behavior change.
