+++ controller/volc_task_delete_test.go
@@ -1,4 +1,4 @@
-package graylight
+package controller
 
 import (
 	"net/http"
