+++ router/video-router.go
@@ -2,7 +2,6 @@ package router
 
 import (
 	"github.com/QuantumNous/new-api/controller"
-	controllergray "github.com/QuantumNous/new-api/controller/graylight"
 	"github.com/QuantumNous/new-api/middleware"
 	relayconstant "github.com/QuantumNous/new-api/relay/constant"
 	"github.com/QuantumNous/new-api/types"
@@ -75,7 +74,7 @@ func SetVideoRouter(router *gin.Engine) {
 		})
 
 		// Task delete: cancel task upstream and refund quota.
-		volcV3Router.DELETE("/contents/generations/tasks/:id", controllergray.VolcTaskDelete)
+		volcV3Router.DELETE("/contents/generations/tasks/:id", controller.VolcTaskDelete)
 	}
 
 	// Jimeng official API routes - direct mapping to official API format
